requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.3 
dj-database-url>=0.5.0
psycopg2-binary==2.9.9
drf-yasg==1.21.7